Add beginning of work to have navbar reflect ethereum account info.
Works for account and ether balance but not weev balance yet.
Having trouble for executing contract calls. It might work how I
have it but metamask might be messing it up because it apparently
is known to have some problems when using a private chain like I'm using
for development (ganache). It rejects all transactions I try to execute
because it thinks the nonces don't match.