Add improvements on trying to get the balance of weev.
We are now getting a number/undefined, but pretty sure both
of those are wrong.

Also started trying to give test tokens but doesn't work yet. Or
it might work but we don't know because get balance probably doesn't.

Unsure if it's a problem with how we're calling the contract functions
or if it's something weird with asynch web3 webapp stuff.